==> notebooks/01-get_data.ipynb <==
import sys
import subprocess
import pkg_resources
import os

sys.path.append(os.path.abspath('../scripts'))

from myFunctions import install_packages, save_table 
install_packages()

import pandas as pd
import numpy as np
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
pd.set_option('display.max_columns', None)


import MetaTrader5 as mt5
import pandas as pd
from datetime import datetime
import MetaTrader5 as mt5
from datetime import datetime
import os

def connect_to_mt5(login: int = None,
                   server: str = "YourServer", 
                   password: str = "YourPassword") -> bool:
    """
    Connects to MetaTrader 5 using the provided login, server, and password.
    
    Parameters:
    - login: MetaTrader login (default: your provided login).
    - server: MetaTrader server (default: your provided server).
    - password: MetaTrader password (default: your provided password).
    """
    if not mt5.initialize(login=login, password=password, server=server):
        print("Failed to connect to MetaTrader 5")
    else:
        print("Successfully connected to MetaTrader 5")
  
def download_and_save_from_mt5(symbols: list = None,
                               timeframe = mt5.TIMEFRAME_M15,
                               start_date: datetime = None,
                               end_date: datetime = datetime.now(),
                               save_dir: str = None):
                            
    """
    Downloads historical data for multiple assets from MetaTrader 5 and saves the data as CSV and Parquet files.
    
    Parameters:
    - symbols: List of asset symbols to download.
    - timeframe: Timeframe for the data (e.g., M15, M1, H1, etc.).
    - start_date: Start date for historical data.
    - end_date: End date for historical data.
    - save_dir: Directory to save the CSV and Parquet files.
    - login: MetaTrader login.
    - server: MetaTrader server.
    - password: MetaTrader password.
    """

    if not os.path.exists(save_dir):
        os.makedirs(save_dir)

    for symbol in symbols:
        
        try:
            rates = mt5.copy_rates_range(symbol, timeframe, start_date, end_date)
            if rates is None or len(rates) == 0:
                print(f"No data found for asset: {symbol}")
                continue

            df = pd.DataFrame(rates)
            df['time'] = pd.to_datetime(df['time'], unit='s')
            first_date = df['time'].iloc[0].strftime('%Y%m%d')
            filename_csv = f"{save_dir}/{symbol}_data_{first_date}.csv"
            filename_parquet = f"{save_dir}/{symbol}_data_{first_date}.parquet"
            df.to_csv(filename_csv, index=False)
            df.to_parquet(filename_parquet, index=False)
            print(f"Data for {symbol} saved to {filename_csv} and {filename_parquet}")

        except Exception as e:
            print(f"Error downloading data for {symbol}: {e}")

    mt5.shutdown()

#Funções do script
def connect_to_mt5(login: int = None,
                   server: str = "YourServer", 
                   password: str = "YourPassword") -> bool:
    """
    Connects to MetaTrader 5 using the provided login, server, and password.
    
    Parameters:
    - login: MetaTrader login (default: your provided login).
    - server: MetaTrader server (default: your provided server).
    - password: MetaTrader password (default: your provided password).
    """
    if not mt5.initialize(login=login, password=password, server=server):
        print("Failed to connect to MetaTrader 5")
    else:
        print("Successfully connected to MetaTrader 5")
  
def download_and_save_from_mt5(symbols: list = None,
                               timeframe = mt5.TIMEFRAME_M15,
                               start_date: datetime = None,
                               end_date: datetime = datetime.now(),
                               save_dir: str = None):
                            
    """
    Downloads historical data for multiple assets from MetaTrader 5 and saves the data as CSV and Parquet files.
    
    Parameters:
    - symbols: List of asset symbols to download.
    - timeframe: Timeframe for the data (e.g., M15, M1, H1, etc.).
    - start_date: Start date for historical data.
    - end_date: End date for historical data.
    - save_dir: Directory to save the CSV and Parquet files.
    - login: MetaTrader login.
    - server: MetaTrader server.
    - password: MetaTrader password.
    """

    if not os.path.exists(save_dir):
        os.makedirs(save_dir)

    for symbol in symbols:
        
        try:
            rates = mt5.copy_rates_range(symbol, timeframe, start_date, end_date)
            if rates is None or len(rates) == 0:
                print(f"No data found for asset: {symbol}")
                continue

            df = pd.DataFrame(rates)
            df['time'] = pd.to_datetime(df['time'], unit='s')
            first_date = df['time'].iloc[0].strftime('%Y%m%d')
            filename_csv = os.path.join(save_dir, f"{symbol}_data_{first_date}.csv")
            filename_parquet = os.path.join(save_dir, f"{symbol}_data_{first_date}.parquet")
            df.to_csv(filename_csv, index=False)
            df.to_parquet(filename_parquet, index=False)
            print(f"Data for {symbol} saved to {filename_csv} and {filename_parquet}")

        except Exception as e:
            print(f"Error downloading data for {symbol}: {e}")

    mt5.shutdown()

# List of assets
assets = [
    'AGFS',    # Iagro
    'BGI$',    # Future Catle
    'IND$',    # Future bovespa index
    'IBOV',    # Bovespa Index
    'ICON',    # Consumer Index
    'ICF$',    # Future Coffee
    'CCM$',    # Future Corn
    'IVVB11',  # ETF (S&P500)
    'GOLD11',  # Future Gold
    'DOL$',    # Future Dollar
    'DI1$',    # Future Selic Proxy
    'IFIX'     # Real state index 

]
# Directory to save the data
start_date = datetime(2022,6,1)
base_dir = os.path.join('..', 'data', 'assets')
login = 1016642903
server = "ClearInvestimentos-CLEAR"
password = "Kali$$123$$" 
        

connect_to_mt5(login, server, password)
download_and_save_from_mt5(symbols=assets, 
                           timeframe= mt5.TIMEFRAME_M15,
                           start_date = start_date,
                           save_dir =base_dir )
